Refactor FXML files to point to presenters
The fx:controller property pointed to view implementations. Now it
points to presenters. However, FXML properties are still injected into
views thanks to ViewAssembler.

Guice creates presenters based on modules. So, the purpose of
fx:controller became so Guice knows which presenter to create and inject
dependencies into.

ViewAssembler is a controller factory that returns views and that's why
FXML properties are still being injected to views by FXMLLoader.