Fix main.fxml bug
When manually changing AnchorPane to BorderPane, I left anchor pane
properties on nodes which caused application to crash. It took me some
time to figure this out because SceneBuilder loaded it nicely so I
thought there was a bug in my code. Well, there was a bug in my code,
but thanks to helpful JavaFX error messages it was harder to detect.