Change unit conversion system
Originally, I created a triad method for unit conversion. The parameters
were source value, source unit, and output unit. However, I believe that
is not a clean approach.

The current unit conversion looks like this:

    unitConverter.convert(100, METER, KILOMETER);

If you don't know the order of parameters, you can only guess what this
does. Because of this, I decided to subclass unit converter class for
each category, so the current unit conversion also looks like this:

   lengthUnitConverter.toKilometers(100, METER);

However, this approach also sucks because when units are modified, this
class needs to be modified as well. I want to write code that applies
changes as "automatically" as possible. That's why I implemented the new
unit conversion system that looks like this:

   unitConverter.convert(100, METER).to(KILOMETER);

As you can see, it is clear what is being converted. Perhaps the con of
this approach is using inner class, but I don't really care since the
pros outweight the cons. Although, I must admit I don't know if inner
classes are code smell or not, but they generally look ugly to me.