Remove all tests
I refused to write tests until I settled for software architecture that
I'd consider clean. Because of that, these tests became outdated so now
it's time to rewrite them.

I used to update tests when I changed the architecture, but it became
useless since I'd always have to add, remove or change tests when I'd
change the architecture.

I've developed an architecture I like, so now it's time to create new
tests.