Separate language and number format setting
When end-user selected a language, a number format used in that locale
would be applied across the application along with the language. This
setting has now been split into two settings:
    1. Language
    2. Number Format
Because of this, the end-user can now select any language and number
format they want to use. No automatic number format switching is
performed upon changing a language.