Sort units based on category and base unit factor
Order of units displayed in the user interfaces used to depend on the
enum declaration order of categories and units. However, it's better
not to depend on my brain to sort these enum values, but rather on the
actual values. Also, if I ever happen to change these enums to regular
classes, that transition would be much smoother.

So, unit categories are sorted based on their order index, first. After
that, the list of units mapped to each category is retrieved. Retrieved
lists are sorted based on the base unit factor (the smaller the factor,
the smaller the unit), each list is divided into two, and then added to
its view.